requirements.txt
apache-airflow==2.6.0
requests==2.28.1
kafka-python==2.0.2
cassandra-driver==3.25.0
